requirements.txt
openai==1.45.1
python-dotenv==1.0.0